Updated playground and playground oscillator
