--- Playgrounds/AKTest.playground/contents.xcplayground
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='5.0' target-platform='ios' buildActiveScheme='true' executeOnSourceChanges='true' importAppTypes='true'>
-    <timeline fileName='timeline.xctimeline'/>
-</playground>
