+++ Sources/AudioKit/Nodes/Generators/PlaygroundOscillator.swift
@@ -11,20 +11,22 @@ public class PlaygroundOscillator: Node {
     fileprivate lazy var sourceNode = AVAudioSourceNode { [self] _, _, frameCount, audioBufferList in
         let ablPointer = UnsafeMutableAudioBufferListPointer(audioBufferList)
                 
-        let phaseIncrement = (twoPi / Float(Settings.sampleRate)) * self.frequency
-        for frame in 0..<Int(frameCount) {
-            // Get signal value for this frame at time.
-            let index = Int(self.currentPhase / twoPi * Float(self.waveform!.count))
-            let value = self.waveform![index] * self.amplitude
+        if isStarted {
+            let phaseIncrement = (twoPi / Float(Settings.sampleRate)) * self.frequency
+            for frame in 0..<Int(frameCount) {
+                // Get signal value for this frame at time.
+                let index = Int(self.currentPhase / twoPi * Float(self.waveform!.count))
+                let value = self.waveform![index] * self.amplitude
 
-            // Advance the phase for the next frame.
-            self.currentPhase += phaseIncrement
-            if self.currentPhase >= twoPi { self.currentPhase -= twoPi }
-            if self.currentPhase < 0.0 { self.currentPhase += twoPi }
-            // Set the same value on all channels (due to the inputFormat we have only 1 channel though).
-            for buffer in ablPointer {
-                let buf: UnsafeMutableBufferPointer<Float> = UnsafeMutableBufferPointer(buffer)
-                buf[frame] = value
+                // Advance the phase for the next frame.
+                self.currentPhase += phaseIncrement
+                if self.currentPhase >= twoPi { self.currentPhase -= twoPi }
+                if self.currentPhase < 0.0 { self.currentPhase += twoPi }
+                // Set the same value on all channels (due to the inputFormat we have only 1 channel though).
+                for buffer in ablPointer {
+                    let buf: UnsafeMutableBufferPointer<Float> = UnsafeMutableBufferPointer(buffer)
+                    buf[frame] = value
+                }
             }
         }
         return noErr
@@ -55,5 +57,7 @@ public class PlaygroundOscillator: Node {
         self.waveform = waveform
         self.frequency = frequency
         self.amplitude = amplitude
+        
+        stop()
     }
 }
